requirements.txt
anthropic
python-dotenv
pydantic
scikit-learn
tenacity
numpy